change root component constructor
