+++ src/containers/App.js
@@ -1,5 +1,10 @@
 import React, { Component } from 'react';
 import { Route } from 'react-router-dom';
+
+import 'bootstrap/dist/css/bootstrap.css';
+import 'react-s-alert/dist/s-alert-default.css';
+import 'react-s-alert/dist/s-alert-css-effects/jelly.css';
+
 import Alert from "react-s-alert";
 
 import { PlanTemplateListContainer, NewPlanTemplate, PlanTemplateContainer } from ".";
