+++ src/containers/index.js
@@ -1,9 +1,11 @@
+import Root from './Root';
 import App from './App';
 import NewPlanTemplate from './NewPlanTemplate';
 import PlanTemplateListContainer from './PlanTemplateListContainer';
 import PlanTemplateContainer from './PlanTemplateContainer';
 
 export { 
+    Root,
     App,
     NewPlanTemplate,
     PlanTemplateListContainer,
